client/src/view/ Login y SingUp  CSS
Se da color negro al texto dentro del input y se da color
